+++ src/main/scala/mimir/exec/sqlite/VGTermFunctions.scala
@@ -1,4 +1,4 @@
-package mimir.backend.sqlite
+package mimir.exec.sqlite
 
 import java.sql.SQLException
 import com.typesafe.scalalogging.slf4j.LazyLogging
